参考项目/Django项目/Mobileshop/requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
djangorestframework==3.7.3
mysqlclient==1.3.12
pytz==2017.3
yapf==0.20.0
